fix: move gemini stream field merges to provider
